+++ bin/camouflage.js
@@ -1,72 +1,68 @@
 #!/usr/bin/env node
 var argv = require("yargs").argv;
+const yaml = require("js-yaml");
+const winston = require("winston");
+const fs = require("fs");
+var config = argv.c || argv.config;
 var help = argv.h || argv.help;
-var mocks = argv.m || argv.mocks || process.cwd();
-var port = argv.p || argv.port || 8080;
-var httpsPort = argv.x || argv.secureport || 8443;
-var enableHttps = argv.s || argv.secure || false;
-var key = argv.k || argv.key;
-var cert = argv.c || argv.cert;
-var numCPUs = argv.n || argv.cpus || 1;
 const osCPUs = require("os").cpus().length;
-var info = require("./../package.json");
-const camouflage = require("../dist/index");
 if (help) {
   console.log(
     [
-      "Camouflage v" + info.version,
-      "",
-      "Usage:",
-      "  camouflage",
-      "",
-      "Optional Parameters:",
-      "  -m, --mocks            - Path to mock files",
-      "  -p, --port             - HTTP Port to listen on",
-      "  -x, --secureport       - HTTPS Port to listen on",
-      "  -s, --secure           - include https server if required",
-      "  -k, --key              - server.key file if -s/--secure is set to true",
-      "  -c, --cert             - server.key file if -s/--secure is set to true",
-      "  -n, --cpus             - number of CPUs you want Camouflage to utilize",
-      "",
-      "Example:",
-      "  camouflage -m './mocks' -s -k ./certs/server.key -c ./certs/server.cert",
+      "Create a config.yml file as shown in the sample yml below",
+      `loglevel: 'info'`,
+      `cpus: 1`,
+      `protocols:`,
+      ` http:`,
+      `   mocks_dir: "./mocks"`,
+      `   port: 8080"`,
+      ` https:`,
+      `   enable: false`,
+      `   port: 8443`,
+      `   cert: "./certs/server.cert"`,
+      `   key: "./certs/server.key"`,
+      ` grpc:`,
+      `   enable: false`,
+      `   port: 5000`,
+      `   mocks_dir: "./grpc/mocks"`,
+      `   protos_dir: "./grpc/protos"`,
     ].join("\n")
   );
   process.exit(1);
 }
-if (numCPUs > osCPUs) {
-  console.log("Number of CPUs specified is greater than or equal to availale CPUs. Please specify a lesser number.");
+if (!config) {
+  logger.error("Please provide a config file.");
   process.exit(1);
 }
-if (enableHttps) {
-  if (!key || !cert) {
-    console.log(
-      [
-        "Missing: Mocks Key or Certificate for https server",
-        "Camouflage v" + info.version,
-        "",
-        "Usage:",
-        "  camouflage",
-        "",
-        "Required Parameter:",
-        "",
-        "Optional Parameters:",
-        "  -m, --mocks            - Path to mock files",
-        "  -p, --port             - HTTP Port to listen on",
-        "  -x, --secureport       - HTTPS Port to listen on",
-        "  -s, --secure           - include https server if required",
-        "  -k, --key              - server.key file if -s/--secure is set to true",
-        "  -c, --cert             - server.key file if -s/--secure is set to true",
-        "  -n, --cpus             - number of CPUs you want Camouflage to utilize",
-        "",
-        "Example:",
-        "  camouflage -m './mocks' -s -k ./certs/server.key -c ./certs/server.cert",
-      ].join("\n")
-    );
-  } else {
-    camouflage.start(mocks, port, enableHttps, numCPUs, key, cert, httpsPort);
-  }
-} else {
-  camouflage.start(mocks, port, enableHttps, numCPUs);
+config = yaml.load(fs.readFileSync(config, "utf-8"));
+const logger = winston.createLogger({
+  level: config.loglevel,
+  format: winston.format.combine(
+    winston.format.colorize(),
+    winston.format.simple(),
+    winston.format.timestamp({ format: "YYYY-MM-DD HH:mm:ss" }),
+    winston.format.printf((log) => `${log.timestamp} ${log.level}: ${log.message}` + (log.splat !== undefined ? `${log.splat}` : " "))
+  ),
+  transports: [new winston.transports.Console()],
+});
+let inputs = [
+  config.protocols.http.mocks_dir,
+  config.protocols.http.port,
+  config.protocols.https.enable,
+  config.protocols.grpc.enable,
+  config.cpus,
+  config.protocols.https.key,
+  config.protocols.https.cert,
+  config.protocols.https.port,
+  config.protocols.grpc.port,
+  config.protocols.grpc.mocks_dir,
+  config.protocols.grpc.protos_dir,
+  config.loglevel,
+];
+if (config.cpus > osCPUs) {
+  logger.error("Number of CPUs specified is greater than or equal to availale CPUs. Please specify a lesser number.");
+  process.exit(1);
 }
+const camouflage = require("../dist/index");
+camouflage.start(...inputs);